cr comments

+++ lfric_macros/release_lfric.py
@@ -23,6 +23,7 @@ import re
 import socket
 import subprocess
 import shutil
+import shlex
 
 from apply_macros import (
     ApplyMacros,
@@ -55,7 +56,7 @@ def run_command(command, timelimit=120):
         - result object from subprocess.run
     """
     result = subprocess.run(
-        command.split(),
+        shlex.split(command),
         capture_output=True,
         text=True,
         timeout=timelimit,
@@ -82,10 +83,10 @@ def set_dependency_path(args):
     LFRic Core source
     """
 
-    print("[INFO] Updating dependencies.sh Core source")
+    print("[INFO] Updating dependencies.yaml Core source")
 
     hostname = socket.gethostname()
-    dep_path = os.path.join(args.apps, "dependencies.sh")
+    dep_path = os.path.join(args.apps, "dependencies.yaml")
     with open(dep_path) as f:
         lines = f.readlines()
     in_core = False
@@ -93,11 +94,11 @@ def set_dependency_path(args):
         if line.strip().startswith("lfric_core"):
             in_core = True
         elif in_core and "source:" in line:
-            line = line.split("source:")
-            line = f"{line[0]}source:{hostname}:{os.path.abspath(args.core)}\n"
+            prefix, _, _ = line.partition("source:")
+            line = f"{prefix}source:{hostname}:{os.path.abspath(args.core)}\n"
         elif in_core and "ref:" in line:
-            line = line.split("ref:")
-            line = f"{line[0]}ref:"
+            prefix, _, _ = line.partition("ref:")
+            line = f"{prefix}ref:"
         elif in_core:
             break
         lines[i] = line
@@ -223,7 +224,14 @@ def copy_head_meta(meta_dirs, args):
         head = os.path.join(meta_dir, "HEAD")
         new = os.path.join(meta_dir, args.version)
         shutil.copytree(head, new)
-        command = f"git add {new}"
+        if args.core in new:
+            new = new.removeprefix(args.core)
+            new = new.lstrip("/")
+            command = f"git -C {args.core} add {new}"
+        elif args.apps in new:
+            new = new.removeprefix(args.apps)
+            new = new.lstrip("/")
+            command = f"git -C {args.apps} add {new}"
         _ = run_command(command)
 
 
@@ -271,8 +279,17 @@ def copy_versions_files(meta_dirs, args):
     for meta_dir in meta_dirs:
         versions_file = os.path.join(meta_dir, "versions.py")
         upgrade_file = os.path.join(meta_dir, upgrade_name)
+        if not os.path.exists(versions_file):
+            raise FileNotFoundError(f"The file {versions_file} doesn't exist")
         shutil.copyfile(versions_file, upgrade_file)
-        command = f"git add {upgrade_file}"
+        if args.core in upgrade_file:
+            upgrade_file = upgrade_file.removeprefix(args.core)
+            upgrade_file = upgrade_file.lstrip("/")
+            command = f"git -C {args.core} add {upgrade_file}"
+        elif args.apps in upgrade_file:
+            upgrade_file = upgrade_file.removeprefix(args.apps)
+            upgrade_file = upgrade_file.lstrip("/")
+            command = f"git -C {args.apps} add {upgrade_file}"
         _ = run_command(command)
 
     return upgrade_name
